Adding my first ever, crude attempt at pathfinding
This code was one of the first fun side projects I did as a programmer. I had very little knowledge about programing and none about pathfinding. This is an algorithm I came up with completely from scratch and on my own. I will note that it doesn't always work correctly; as in there definitely exists a path but it cannot be found. Regardless, I had fun and this is just a small piece of my history as a programmer.